Small enhancements
Small improvements

- Game now shows the current version on the settings page
- Game will warn users more frequently, for example when accidentally swiping or pressing the back button
- Other layout changes
- Slightly changed the translations
