+++ SECURITY.md
@@ -6,19 +6,20 @@ We take the security of our software seriously. Below are the versions of Half-a
 
 | Version                                                                                                | Supported          |
 |--------------------------------------------------------------------------------------------------------|--------------------|
-| [V2.6](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/StableV2.6) (current) | :white_check_mark: |
+| [V2.7](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/StableV2.7) (current) | :white_check_mark: |
+| [V2.6](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/StableV2.6)           | :white_check_mark: |
 | [V2.4](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/DebugV2.4)            | :x:                |
 | [V2.3](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/DebugV2.3)            | :x:                |
 | < [V2.3](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/releases/tag/DebugV2.3)          | :x:                |
 
 ## Reporting a Vulnerability
 
-We encourage users to report any security vulnerabilities they find in Half-a-Minute-IT-Madness-Android. To report a vulnerability, please follow these steps:
+We encourage users to report any bugs, feature requests and security vulnerabilities they find in Half-a-Minute-IT-Madness-Android. To report a vulnerability, please follow these steps:
 
-1. **Contact**: Reach out to us via [GitHub Issues](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/issues) or directly by [email](mailto:Github@sten-tijhuis.nl).
+1. **Contact**: Reach out to us via [GitHub Issues](https://github.com/Stensel8/Half-a-Minute-IT-Madness-Android/issues), [the website](http://hmitm.sten-tijhuis.nl) or directly by [email](mailto:Github@sten-tijhuis.nl).
 2. **Describe**: Clearly describe the vulnerability, including steps to reproduce it, and, if possible, a patch or solution.
-3. **Response Time**: We aim to respond to vulnerability reports as quickly as possible, typically within 48 hours.
+3. **Response Time**: Despite our busy student schedules, we're dedicated to addressing vulnerability reports promptly. Our goal is to respond within a week, understanding our limited availability. Your patience is valued.
 4. **Follow-up**: Once a report is received, we will follow up with you to discuss the impact and work towards a resolution.
-5. Disclosure: We will keep you informed about the progress of addressing the vulnerability. Once fixed, a public announcement will be made, crediting your contribution unless you prefer to remain anonymous.
+5. Disclosure: We will keep you informed about the progress of addressing the vulnerability. 
 
 We appreciate your effort to improve the security of Half-a-Minute-IT-Madness-Android and encourage responsible disclosure practices.
