--- app/build.gradle.bak
@@ -1,83 +1,0 @@
-// Apply necessary plugins
-apply plugin: 'com.android.application'
-apply plugin: 'com.google.gms.google-services'
-apply plugin: 'kotlin-android'
-
-android {
-    // Configuration for the Android application
-    defaultConfig {
-        applicationId "com.halfminute.itmadness" // Unique identifier for the app
-        minSdk 30 // Minimum SDK version required
-        targetSdkVersion 34 // Target SDK version
-        compileSdk 34 // Compile SDK version
-        versionCode 26 // Version code for the app (increment for each release)
-        versionName '2.6' // Version name for the app
-        multiDexEnabled true // Enable MultiDex for applications with many methods
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner" // Test runner for instrumentation tests
-        signingConfig signingConfigs.debug // Signing configuration (debug build)
-    }
-
-    // Enable view binding
-    buildFeatures {
-        viewBinding true
-    }
-
-    // Build types configuration
-    buildTypes {
-        release {
-            minifyEnabled false // Disable code shrinking for release builds
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro' // ProGuard configuration
-        }
-    }
-
-    // Configure Java compilation options
-    android {
-        compileOptions {
-            sourceCompatibility JavaVersion.VERSION_17
-            targetCompatibility JavaVersion.VERSION_17
-        }
-
-        // Enable lint checks and deprecation warnings for Java compilation
-        tasks.withType(JavaCompile).tap {
-            configureEach {
-                options.compilerArgs += ['-Xlint:all', '-Xlint:deprecation']
-            }
-        }
-    }
-
-    // Namespace for the Android application
-    namespace 'com.halfminute.itmadness'
-}
-
-// Dependencies for the Android application
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'com.google.android.material:material:1.10.0'
-    implementation 'com.google.code.gson:gson:2.10.1'
-    implementation 'com.codesgood:justifiedtextview:2.0.1'
-    implementation 'com.google.firebase:firebase-analytics:21.5.0'
-    implementation 'com.google.firebase:firebase-auth:22.3.0'
-    implementation 'com.google.firebase:firebase-firestore:24.9.1'
-    implementation 'androidx.core:core-ktx:1.12.0'
-    implementation "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-}
-
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-buildscript {
-    repositories {
-        google()
-        mavenCentral()
-    }
-
-    dependencies {
-        // Gradle build tool dependencies
-        classpath 'com.android.tools.build:gradle:8.2.1'
-        classpath 'com.google.gms:google-services:4.4.0'
-    }
-}
