--- app/src/main/res/drawable/border.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape= "rectangle">
-    <solid android:color="?attr/backgroundColor"/>
-    <stroke android:width="1dp" android:color="?attr/textColor"/>
-</shape>
