--- app/src/main/res/drawable/list_divider.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
-    <solid android:color="?attr/textColor" />
-    <corners android:radius="2dp"/>
-</shape>
