+++ app/src/main/res/layout/choose_game.xml
@@ -64,7 +64,6 @@
 
         <!-- Button start guessing game -->
         <Button
-            android:id="@+id/guessingGameButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
