--- app/src/main/res/layout/listrow.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:textColor="?attr/textColor"
-    android:textSize="24sp"
-
-    />
