+++ app/src/main/res/layout/settings.xml
@@ -157,7 +157,7 @@
             android:textColor="@color/buttonTextColor"
             android:textSize="16sp" />
 
-        <!-- Welcome and Updates Button -->
+        <!-- Welcome and Releasenotes Button -->
         <Button
             android:id="@+id/btnShowWelcomeAndUpdates"
             android:layout_width="200dp"
@@ -170,5 +170,62 @@
             android:textColor="@color/buttonTextColor"
             android:textSize="16sp" />
 
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="20dp"
+            android:gravity="center">
+
+            <!-- Logo -->
+            <ImageView
+                android:id="@+id/imgGroupLogo"
+                android:layout_width="219dp"
+                android:layout_height="158dp"
+                android:layout_marginBottom="-1dp"
+                android:contentDescription="@string/group_logo_description"
+                android:src="@drawable/beunhazen_logo_transparent" />
+
+            <!-- Group Name -->
+            <TextView
+                android:id="@+id/tvGroupName"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/imgGroupLogo"
+                android:layout_gravity="center_horizontal"
+                android:layout_marginTop="-55dp"
+                android:layout_marginStart="60dp"
+                android:fontFamily="@font/segoeui"
+                android:text="@string/groupname"
+                android:textColor="?attr/textColor"
+                android:textSize="16sp" />
+
+
+            <!-- Build Info -->
+            <TextView
+                android:id="@+id/tvBuildInfo"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/tvGroupName"
+                android:layout_marginTop="30dp"
+                android:layout_marginStart="-50dp"
+                android:fontFamily="@font/segoeui"
+                android:text="@string/build_label"
+                android:textColor="?attr/textColor"
+                android:textSize="16sp" />
+
+            <!-- Version Info -->
+            <TextView
+                android:id="@+id/tvVersionInfo"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/tvBuildInfo"
+                android:layout_marginTop="1dp"
+                android:layout_marginStart="-50dp"
+                android:fontFamily="@font/segoeui"
+                android:text="@string/version_label"
+                android:textColor="?attr/textColor"
+                android:textSize="16sp" />
+
+        </RelativeLayout>
     </LinearLayout>
 </ScrollView>
