--- app/src/main/res/layout/welcome_message.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="?attr/backgroundColor"
-    android:orientation="vertical"
-    android:padding="5dp">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <!-- TextView welcome (Header) -->
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginTop="0dp"
-                android:fontFamily="@font/segoeui"
-                android:text="@string/welcome"
-                android:textColor="?attr/textColor"
-                android:textSize="36sp" />
-
-            <!-- Imagebutton close button -->
-            <ImageButton
-                android:id="@+id/btnClose"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_gravity="center_vertical|end"
-                android:background="?attr/backgroundColor"
-                android:contentDescription="@string/closeWindow"
-                android:minWidth="48dp"
-                app:srcCompat="@android:drawable/ic_menu_close_clear_cancel" />
-
-        </FrameLayout>
-
-        <!-- TextView welcome text -->
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:fontFamily="@font/segoeui"
-            android:text="@string/welcomeText"
-            android:textAllCaps="false"
-            android:textColor="?attr/textColor"
-            android:textSize="18sp" />
-
-        <Button
-            android:id="@+id/btnViewReleaseNotes"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="10dp"
-            android:text="@string/show_releasenotes" />
-
-    </LinearLayout>
-</ScrollView>
