+++ app/src/main/res/values-fr-rFR/strings.xml
@@ -1,59 +1,78 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <!-- Game Buttons -->
+    <string name="playGame">Jouer au jeu</string>
+    <string name="mathGameButton">Jeu de mathématiques</string>
+    <string name="languageGameButton">Test de langues</string>
+    <string name="guessingGameButton">Jeu de devinette</string>
+    <string name="defaultBtnValue">3</string>
+
+    <!-- Titles -->
     <string name="app_name">Half a Minute IT Madness</string>
-    <string name="FrenchToEnglish">Anglais à Français</string>
-    <string name="PersonalHighScore">Meilleur score : </string>
-    <string name="DutchToEnglish">Néerlandais à Français</string>
-    <string name="guessingGameButton">Jeu de devinettes</string>
+    <string name="title_activity_language_test">Test de langues</string>
     <string name="chooseGameTitle">Choisissez votre jeu</string>
     <string name="chooseTestTitle">Choisissez votre test</string>
-    <string name="darkModeText">Mode sombre</string>
-    <string name="difficultyEasy">Facile</string>
-    <string name="difficultyHard">Difficile</string>
-    <string name="difficultyMedium">Moyen</string>
     <string name="difficultyTitle">Difficulté</string>
-    <string name="EnglishToFrench">Français à Anglais</string>
-    <string name="exitGame">Quitter</string>
-    <string name="languageGameButton">Test de langues</string>
-    <string name="mainMenuBtn">Menu principal</string>
-    <string name="mathGameButton">Jeu de math</string>
     <string name="pauseMenuTitle">Menu pause</string>
-    <string name="playGame">Jouer</string>
-    <string name="resumeBtn">Rejouer</string>
-    <string name="retry">Réessayer</string>
+    <string name="gameTitle">Titre du jeu</string>
+
+    <!-- Settings -->
     <string name="settingsTitle">Paramètres</string>
-    <string name="ScoreValue">0/0</string>
-    <string name="TimerValue">30s</string>
-    <string name="defaultAnswerText">\?</string>
-    <string name="defaultBtnValue">3</string>
-    <string name="hScoreDescription">High-Score image</string>
-    <string name="scoreGOver">"Score: "</string>
-    <string name="ChangeLanguage">Changer la langue</string>
-    <string name="correct">Correct!</string>
-    <string name="wrong">Faux!</string>
-    <string name="question">Ik ben 23 jaar oud</string>
-    <string name="sum">1 + 2 =</string>
-    <string name="title_activity_language_test">Language_test</string>
-    <string name="answerLang">I am 23 years old</string>
-    <string name="EnglishToDutch">Français à Néerlandais</string>
-    <string name="changeGame">Changer de jeu</string>
-    <string name="changeDifficultyTo">"Difficulté: "</string>
-    <string name="changeDfficultyTo">Top 5:</string>
-    <string name="chosenGame">"Jeu: "</string>
-    <string name="or">Ou</string>
-    <string name="no_data">Aucun résultat</string>
-    <string name="welcome">Bienvenue!</string>
-    <string name="welcomeText">
-    <![CDATA[
-    "Découvrez Half a Minute IT Madness, un jeu créatif conçu par trois étudiants dévoués de la Saxion University of Applied Sciences. Choisissez parmi trois jeux stimulants : mathématiques, langues et devinettes. Chaque jeu dure 30 secondes, où une réponse correcte rapporte un point et une réponse incorrecte coûte une vie. Surveillez vos vies, affichées comme des cœurs vert/bleu en haut à droite. Le jeu se termine après 30 secondes ou en perdant toutes les vies. Challengez-vous pour améliorer votre score !
-    ]]>
-    </string>
-    <string name="closeWindow">Fermer la fenêtre</string>
-    <string name="yes">Oui</string>
-    <string name="no">Non</string>
+    <string name="darkModeText">Mode sombre</string>
     <string name="sound">Son</string>
-    <string name="GermanToEnglish">German To English</string>
-    <string name="EnglishToGerman">English to German</string>
-    <string name="show_releasenotes">Voir les notes de version</string>
-    <string name="guessing_game">Guessing Game</string>
+    <string name="ChangeLanguage">Changer de langue</string>
+    <string name="show_releasenotes">Afficher les notes de version</string>
+
+    <!-- Timer and Score -->
+    <string name="TimerValue">30s</string>
+    <string name="game_time">30 secondes de jeu</string>
+    <string name="ScoreValue">0/0</string>
+
+    <!-- Difficulty Levels -->
+    <string name="difficultyEasy">Facile</string>
+    <string name="difficultyMedium">Moyen</string>
+    <string name="difficultyHard">Difficile</string>
+
+    <!-- Buttons and Labels -->
+    <string name="resumeBtn">Rejouer</string>
+    <string name="mainMenuBtn">Menu principal</string>
+    <string name="retry">Réessayer</string>
+    <string name="exitGame">Quitter le jeu</string>
+    <string name="changeGame">Changer de jeu</string>
+    <string name="changeDifficultyTo">"Difficulté : "</string>
+    <string name="changeDfficultyTo">Top 5 :</string>
+    <string name="chosenGame">"Jeu : "</string>
+
+    <!-- Score Messages -->
+    <string name="scoreGOver">"Score : "</string>
+    <string name="PersonalHighScore">"Record personnel : "</string>
+    <string name="correct">Correct !</string>
+    <string name="wrong">Incorrect !</string>
+
+    <!-- Lives and Questions -->
+    <string name="sum">1 + 2 =</string>
+    <string name="question">J\'ai 23 ans</string>
+    <string name="answerLang">I am 23 years old</string>
+
+    <!-- Welcome and Information -->
+    <string name="welcome">Bienvenue dans Half a Minute IT Madness !</string>
+    <string name="welcomeText">
+        <![CDATA[
+        Découvrez Half a Minute IT Madness, un jeu créatif conçu par trois étudiants dévoués de la Saxion University of Applied Sciences. Choisissez parmi trois jeux stimulants : mathématiques, langues et devinettes. Chaque jeu dure 30 secondes, où une réponse correcte rapporte un point et une réponse incorrecte coûte une vie. Gardez un œil sur vos vies, affichées sous forme de cœurs verts/bleus en haut à droite. Le jeu se termine après 30 secondes ou en perdant toutes les vies. Défiez-vous pour améliorer votre score !
+        ]]>
+    </string>
+
+    <!-- Dialog and Confirmation -->
+    <string name="closeWindow">Fermer la fenêtre</string>
+
+    <!-- Language Options -->
+    <string name="DutchToEnglish">Néerlandais vers l\'anglais</string>
+    <string name="FrenchToEnglish">Français vers l\'anglais</string>
+    <string name="EnglishToFrench">Anglais vers le français</string>
+    <string name="EnglishToDutch">Anglais vers le néerlandais</string>
+    <string name="GermanToEnglish">Allemand vers l\'anglais</string>
+    <string name="EnglishToGerman">Anglais vers l\'allemand</string>
+
+    <!-- Miscellaneous -->
+    <string name="theme_dynamic_icon">Il s\'agit d\'une icône dynamique qui changera en fonction des paramètres du mode sombre</string>
+    <string name="hScoreDescription">Image du meilleur score</string>
 </resources>
