+++ app/src/main/res/values-nl-rNL/strings.xml
@@ -1,60 +1,78 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Half a Minute IT Madness</string>
-    <string name="TimerValue">30s</string>
-    <string name="DutchToEnglish">Engels naar Nederlands</string>
+    <!-- Game Buttons -->
+    <string name="playGame">Spelen</string>
+    <string name="mathGameButton">Rekenspel</string>
+    <string name="languageGameButton">Taaltest</string>
     <string name="guessingGameButton">Raadspel</string>
+    <string name="defaultBtnValue">3</string>
+
+    <!-- Titles -->
+    <string name="app_name">Half a Minute IT Madness</string>
+    <string name="title_activity_language_test">Taaltest</string>
     <string name="chooseGameTitle">Kies je spel</string>
     <string name="chooseTestTitle">Kies je test</string>
+    <string name="difficultyTitle">Moeilijkheid</string>
+    <string name="pauseMenuTitle">Pauzemenu</string>
+    <string name="gameTitle">Speltitel</string>
+
+    <!-- Settings -->
+    <string name="settingsTitle">Instellingen</string>
     <string name="darkModeText">Donkere modus</string>
-    <string name="defaultAnswerText">\?</string>
-    <string name="defaultBtnValue">3</string>
+    <string name="sound">Geluid</string>
+    <string name="ChangeLanguage">Verander taal</string>
+    <string name="show_releasenotes">Toon release-opmerkingen</string>
+
+    <!-- Timer and Score -->
+    <string name="TimerValue">30s</string>
+    <string name="game_time">30 seconden spel</string>
+    <string name="ScoreValue">0/0</string>
+
+    <!-- Difficulty Levels -->
     <string name="difficultyEasy">Makkelijk</string>
     <string name="difficultyMedium">Gemiddeld</string>
     <string name="difficultyHard">Moeilijk</string>
-    <string name="difficultyTitle">Moeilijkheid</string>
-    <string name="EnglishToFrench">Nederlands naar Engels</string>
-    <string name="exitGame">Exit game</string>
-    <string name="hScoreDescription">High-Score image</string>
-    <string name="languageGameButton">Taaltest</string>
+
+    <!-- Buttons and Labels -->
+    <string name="resumeBtn">Opnieuw afspelen</string>
     <string name="mainMenuBtn">Hoofdmenu</string>
-    <string name="pauseMenuTitle">Pauzemenu</string>
-    <string name="playGame">Spelen</string>
-    <string name="resumeBtn">Spelen</string>
     <string name="retry">Opnieuw proberen</string>
-    <string name="scoreGOver">"Score: "</string>
-    <string name="settingsTitle">Instellingen</string>
-    <string name="FrenchToEnglish">Frans naar Nederlands</string>
-    <string name="PersonalHighScore">"High-score: "</string>
-    <string name="ScoreValue">0/0</string>
-    <string name="ChangeLanguage">Taal veranderen</string>
-    <string name="correct">Juist!</string>
-    <string name="wrong">Verkeerd!</string>
-    <string name="question">I am 23 years old</string>
-    <string name="sum">1 + 2 =</string>
-    <string name="title_activity_language_test">Language_test</string>
-    <string name="mathGameButton">Wiskundespel</string>
-    <string name="answerLang">I am 23 years old</string>
-    <string name="EnglishToDutch">Nederlands naar Frans</string>
-    <string name="changeGame">Spel veranderen</string>
+    <string name="exitGame">Spel verlaten</string>
+    <string name="changeGame">Verander spel</string>
     <string name="changeDifficultyTo">"Moeilijkheid: "</string>
     <string name="changeDfficultyTo">Top 5:</string>
     <string name="chosenGame">"Spel: "</string>
-    <string name="or">Of</string>
-    <string name="no_data">Geen resultaat</string>
-    <string name="welcome">Welkom!</string>
-    <string name="welcomeText">
-    <![CDATA[
-    Ontdek Half a Minute IT Madness, een creatieve game ontworpen door drie toegewijde Saxion Hogeschool studenten. Kies uit drie uitdagende spellen: rekenen, taal, en raden. Elk spel duurt 30 seconden, waarbij een correct antwoord een punt oplevert en een fout antwoord een leven kost. Let op je levens, weergegeven als groen/blauw achtige harten rechtsboven. Het spel eindigt na 30 seconden of bij verlies van alle levens. Daag jezelf uit om je score te verbeteren!
-    ]]>
 
+    <!-- Score Messages -->
+    <string name="scoreGOver">"Score: "</string>
+    <string name="PersonalHighScore">"Persoonlijk record: "</string>
+    <string name="correct">Correct!</string>
+    <string name="wrong">Fout!</string>
+
+    <!-- Lives and Questions -->
+    <string name="sum">1 + 2 =</string>
+    <string name="question">Ik ben 23 jaar oud</string>
+    <string name="answerLang">Ik ben 23 jaar oud</string>
+
+    <!-- Welcome and Information -->
+    <string name="welcome">Welkom bij Half a Minute IT Madness!</string>
+    <string name="welcomeText">
+        <![CDATA[
+        Ontdek Half a Minute IT Madness, een creatief spel ontworpen door drie toegewijde studenten van de Saxion University of Applied Sciences. Kies uit drie uitdagende spellen: rekenen, taal en raden. Elk spel duurt 30 seconden, waarbij een juist antwoord een punt oplevert en een fout antwoord een leven kost. Houd je levens in de gaten, weergegeven als groen/blauwe harten rechtsboven. Het spel eindigt na 30 seconden of als alle levens verloren zijn. Daag jezelf uit om je score te verbeteren!
+        ]]>
     </string>
+
+    <!-- Dialog and Confirmation -->
     <string name="closeWindow">Venster sluiten</string>
-    <string name="yes">Ja</string>
-    <string name="no">Nee</string>
-    <string name="sound">Audio</string>
-    <string name="GermanToEnglish">German To English</string>
-    <string name="EnglishToGerman">English to German</string>
-    <string name="show_releasenotes">Toon releasenotes</string>
-    <string name="guessing_game">Guessing Game</string>
+
+    <!-- Language Options -->
+    <string name="DutchToEnglish">Nederlands naar Engels</string>
+    <string name="FrenchToEnglish">Frans naar Engels</string>
+    <string name="EnglishToFrench">Engels naar Frans</string>
+    <string name="EnglishToDutch">Engels naar Nederlands</string>
+    <string name="GermanToEnglish">Duits naar Engels</string>
+    <string name="EnglishToGerman">Engels naar Duits</string>
+
+    <!-- Miscellaneous -->
+    <string name="theme_dynamic_icon">Dit is een dynamische pictogram dat verandert op basis van de instellingen van de donkere modus</string>
+    <string name="hScoreDescription">Highscore-afbeelding</string>
 </resources>
