+++ app/src/main/res/values/strings.xml
@@ -1,67 +1,93 @@
 <resources>
-    <string name="app_name">Half a Minute IT Madness</string>
+    <!-- Game Buttons -->
+    <string name="playGame">Play game</string>
     <string name="mathGameButton">Math Game</string>
-    <string name="languageGameButton">Language test</string>
-    <string name="guessingGameButton">Guessing game</string>
-    <string name="chooseGameTitle">Choose your game</string>
-    <string name="chooseTestTitle">Choose your test</string>
-    <string name="DutchToEnglish">Dutch to English</string>
-    <string name="settingsTitle">Settings</string>
+    <string name="languageGameButton">Language Test</string>
+    <string name="guessingGameButton">Guessing Game</string>
+    <string name="defaultBtnValue">3</string>
+
+    <!-- Titles -->
+    <string name="app_name">Half a Minute IT Madness</string>
+    <string name="title_activity_language_test">Language_test</string>
+    <string name="chooseGameTitle">Choose Your Game</string>
+    <string name="chooseTestTitle">Choose Your Test</string>
     <string name="difficultyTitle">Difficulty</string>
+    <string name="pauseMenuTitle">Pause Menu</string>
+    <string name="gameTitle">Game Title</string>
+
+    <!-- Settings -->
+    <string name="settingsTitle">Settings</string>
+    <string name="darkModeText">Dark Mode</string>
+    <string name="sound">Sound</string>
+    <string name="ChangeLanguage">Change language</string>
+    <string name="show_releasenotes">Show Releasenotes</string>
+
+    <!-- Timer and Score -->
     <string name="TimerValue">30s</string>
-    <string name="game_time" translatable="false">30 seconds game</string>
+    <string name="game_time">30 seconds game</string>
     <string name="timer_seconds" translatable="false">%1$d s</string>
     <string name="ScoreValue">0/0</string>
     <string name="score_format" translatable="false">%1$d/%2$d</string>
+
+    <!-- Difficulty Levels -->
     <string name="difficultyEasy">Easy</string>
     <string name="difficultyMedium">Medium</string>
     <string name="difficultyHard">Hard</string>
-    <string name="darkModeText">Dark mode</string>
-    <string name="pauseMenuTitle">Pause menu</string>
+
+    <!-- Buttons and Labels -->
     <string name="resumeBtn">Replay</string>
-    <string name="mainMenuBtn">Main menu</string>
-    <string name="defaultBtnValue">3</string>
-    <string name="defaultAnswerText">\?</string>
-    <string name="retry">Try again</string>
-    <string name="exitGame">Exit game</string>
-    <string name="scoreGOver">"Score: "</string>
-    <string name="PersonalHighScore">"Personal best: "</string>
-    <string name="hScoreDescription">High-Score image</string>
-    <string name="FrenchToEnglish">French to English</string>
-    <string name="playGame">Play game</string>
-    <string name="ChangeLanguage">Change language</string>
-    <string name="correct">Correct!</string>
-    <string name="wrong">Wrong!</string>
-    <string name="lives" translatable="false">%d</string>
-    <string name="sum">1 + 2 =</string>
-    <string name="title_activity_language_test">Language_test</string>
-    <string name="question">Ik ben 23 jaar oud</string>
-    <string name="points" translatable="false">%d/%d</string>
-    <string name="answerLang">I am 23 years old</string>
-    <string name="EnglishToFrench">English to French</string>
-    <string name="EnglishToDutch">English to dutch</string>
-    <string name="changeGame">Change game</string>
+    <string name="mainMenuBtn">Main Menu</string>
+    <string name="retry">Try Again</string>
+    <string name="exitGame">Exit Game</string>
+    <string name="changeGame">Change Game</string>
     <string name="changeDifficultyTo">"Difficulty: "</string>
     <string name="changeDfficultyTo">Top 5:</string>
     <string name="chosenGame">"Game: "</string>
-    <string name="or">Or</string>
-    <string name="no_data">No result</string>
 
+    <!-- Score Messages -->
+    <string name="scoreGOver">"Score: "</string>
+    <string name="PersonalHighScore">"Personal Best: "</string>
+    <string name="correct">Correct!</string>
+    <string name="wrong">Wrong!</string>
+
+    <!-- Lives and Questions -->
+    <string name="lives" translatable="false">%d</string>
+    <string name="sum">1 + 2 =</string>
+    <string name="question">Ik ben 23 jaar oud</string>
+    <string name="defaultAnswerText" translatable="false">\?</string>
+    <string name="points" translatable="false">%d/%d</string>
+    <string name="answerLang">I am 23 years old</string>
+
+    <!-- Welcome and Information -->
     <string name="welcome">Welcome to Half a Minute IT Madness!</string>
     <string name="welcomeText">
-    <![CDATA[
-    Discover Half a Minute IT Madness, a creative game designed by three dedicated Saxion University of Applied Sciences students. Choose from three challenging games: math, language, and guessing. Each game lasts for 30 seconds, where a correct answer earns a point and a wrong answer costs a life. Keep an eye on your lives, displayed as green/blue hearts in the top right. The game ends after 30 seconds or upon losing all lives. Challenge yourself to improve your score!
-
-    ]]>
+        <![CDATA[
+        Discover Half a Minute IT Madness, a creative game designed by three dedicated Saxion University of Applied Sciences students. Choose from three challenging games: math, language, and guessing. Each game lasts for 30 seconds, where a correct answer earns a point and a wrong answer costs a life. Keep an eye on your lives, displayed as green/blue hearts in the top right. The game ends after 30 seconds or upon losing all lives. Challenge yourself to improve your score!
+        ]]>
     </string>
-    <string name="closeWindow">Close window</string>
-    <string name="gameTitle" translatable="false">Game title</string>
-    <string name="yes">yes</string>
-    <string name="no">No</string>
-    <string name="sound">Sound</string>
-    <string name="GermanToEnglish">German To English</string>
+
+    <!-- Dialog and Confirmation -->
+    <string name="closeWindow">Close Window</string>
+
+    <!-- Language Options -->
+    <string name="DutchToEnglish">Dutch to English</string>
+    <string name="FrenchToEnglish">French to English</string>
+    <string name="EnglishToFrench">English to French</string>
+    <string name="EnglishToDutch">English to Dutch</string>
+    <string name="GermanToEnglish">German to English</string>
     <string name="EnglishToGerman">English to German</string>
-    <string name="show_releasenotes">Show releasenotes</string>
-    <string name="guessing_game">Guessing Game</string>
-    <string name="theme_dynamic_icon" translatable="false">This is an dynamic icon that will change based on the darktheme settings</string>
+
+    <!-- Miscellaneous -->
+    <string name="theme_dynamic_icon">This is a dynamic icon that will change based on the dark theme settings</string>
+    <string name="hScoreDescription">High-Score image</string>
+
+    <!-- Group Information -->
+    <string name="group_logo_description" translatable="false">De Beunhazen</string>
+    <string name="groupname" translatable="false">De Beunhazen</string>
+
+    <!-- Build and Version Information -->
+    <string name="build_label" translatable="false">Build: </string>
+    <string name="version_label" translatable="false">Version: </string>
+    <string name="versioncode" translatable="false">27</string>
+    <string name="versionname" translatable="false">2.7</string>
 </resources>
